feat: Set up hostname
Launching Spyndra required us to look up the IP address, which is
annoying. So I wrote a simple script that change the hostname to be
spyndra.local. If this works out, ssh into Spyndra should be as easy as
a line of command. README file is provided to guide installation and
help people how to use it.